Use spring default json serialization in Perftest controller
