+++ ngrinder-controller/src/test/java/org/ngrinder/perftest/controller/PerfTestControllerTest.java
@@ -33,7 +33,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.HttpEntity;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.ui.ModelMap;
 
@@ -42,6 +41,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.*;
@@ -55,7 +55,10 @@ import static org.junit.Assert.*;
 public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest implements WebConstants {
 
 	@Autowired
-	private MockPerfTestController controller;
+	private MockPerfTestApiController mockPerfTestApiController;
+
+	@Autowired
+	private MockPerfTestController mockPerfTestController;
 
 	@Autowired
 	private Config config;
@@ -74,20 +77,20 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 	@Test
 	public void testGetPerfTestDetail() {
 		ModelMap model = new ModelMap();
-		controller.getOne(getTestUser(), null, model);
+		mockPerfTestController.getOne(getTestUser(), null, model);
 		assertThat(model.get(PARAM_TEST), notNullValue());
 		model.clear();
 
-		controller.getOne(getTestUser(), 0L, model);
+		mockPerfTestController.getOne(getTestUser(), 0L, model);
 		assertThat(model.get(PARAM_TEST), notNullValue());
 		model.clear();
 		long invalidId = 123123123123L;
-		controller.getOne(getTestUser(), invalidId, model);
+		mockPerfTestController.getOne(getTestUser(), invalidId, model);
 		assertThat(model.get(PARAM_TEST), notNullValue());
 
 		PerfTest createPerfTest = createPerfTest("hello", Status.READY, new Date());
 		model.clear();
-		controller.getOne(getTestUser(), createPerfTest.getId(), model);
+		mockPerfTestController.getOne(getTestUser(), createPerfTest.getId(), model);
 		assertThat(model.get(PARAM_TEST), notNullValue());
 
 	}
@@ -98,7 +101,7 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		FileUtils.deleteQuietly(file);
 		CompressionUtils.unzip(new ClassPathResource("TEST_USER.zip").getFile(), file);
 		repo.setUserRepository(new File(file, getTestUser().getUserId()));
-		controller.getResources(getTestUser(), "filefilter.txt", null);
+		mockPerfTestApiController.getResources(getTestUser(), "filefilter.txt", null);
 	}
 
 	@Test
@@ -106,18 +109,18 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		String testName = "test1";
 		PerfTest test = createPerfTest(testName, Status.READY, new Date());
 		ModelMap model = new ModelMap();
-		controller.delete(getTestUser(), String.valueOf(test.getId()));
+		mockPerfTestApiController.delete(getTestUser(), String.valueOf(test.getId()));
 		model.clear();
 		PerfTest test1 = createPerfTest(testName, Status.READY, new Date());
 		PerfTest test2 = createPerfTest(testName, Status.READY, new Date());
 		String delIds = "" + test1.getId() + "," + test2.getId();
-		controller.delete(getTestUser(), delIds);
+		mockPerfTestApiController.delete(getTestUser(), delIds);
 
 		model.clear();
-		controller.getOne(getTestUser(), test1.getId(), model);
+		mockPerfTestController.getOne(getTestUser(), test1.getId(), model);
 		assertThat(((PerfTest) model.get(PARAM_TEST)).getId(), nullValue());
 		model.clear();
-		controller.getOne(getTestUser(), test2.getId(), model);
+		mockPerfTestController.getOne(getTestUser(), test2.getId(), model);
 		assertThat(((PerfTest) model.get(PARAM_TEST)).getId(), nullValue());
 	}
 
@@ -131,20 +134,20 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		cloneTest.setId(test.getId()); // set cloned test's ID as previous test
 
 		ModelMap model = new ModelMap();
-		controller.saveOne(getTestUser(), cloneTest, true, model);
+		mockPerfTestController.saveOne(getTestUser(), cloneTest, true, model);
 		assertThat(preId, not(cloneTest.getId()));
 
 		// test leave comment
-		controller.leaveComment(getTestUser(), cloneTest.getId(), "TestComment", "");
+		mockPerfTestApiController.leaveComment(getTestUser(), cloneTest.getId(), "TestComment", "");
 		model.clear();
-		controller.getOne(getTestUser(), cloneTest.getId(), model);
+		mockPerfTestController.getOne(getTestUser(), cloneTest.getId(), model);
 		PerfTest testInDB = (PerfTest) model.get(PARAM_TEST);
 		assertThat(testInDB.getTestComment(), is("TestComment"));
 
 		// test stop test
 		cloneTest.setStatus(Status.TESTING);
 		perfTestService.save(getTestUser(), cloneTest);
-		controller.stop(getTestUser(), String.valueOf(cloneTest.getId()));
+		mockPerfTestApiController.stop(getTestUser(), String.valueOf(cloneTest.getId()));
 	}
 
 	/**
@@ -174,18 +177,19 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		newTest.setVuserPerAgent(newTest.getProcesses() * newTest.getThreads());
 		newTest.setRegion(config.getRegion());
 		newTest.setAgentCount(1);
+		newTest.setCreatedUser(getTestUser());
 
 		ModelMap model = new ModelMap();
-		controller.saveOne(getTestUser(), newTest, false, model);
-		controller.getOne(getTestUser(), newTest.getId(), model);
+		mockPerfTestController.saveOne(getTestUser(), newTest, false, model);
+		mockPerfTestController.getOne(getTestUser(), newTest.getId(), model);
 		PerfTest testInDB = (PerfTest) model.get(PARAM_TEST);
 		assertThat(testInDB.getTestName(), is(newName));
 		assertThat(testInDB.getId(), is(test.getId()));
 
 		model.clear();
 		newTest.setStatus(Status.READY);
-		controller.saveOne(getTestUser(), newTest, false, model);
-		controller.getOne(getTestUser(), newTest.getId(), model);
+		mockPerfTestController.saveOne(getTestUser(), newTest, false, model);
+		mockPerfTestController.getOne(getTestUser(), newTest.getId(), model);
 		testInDB = (PerfTest) model.get(PARAM_TEST);
 		assertThat(testInDB.getTestName(), is(newName));
 		assertThat(testInDB.getId(), is(test.getId()));
@@ -194,7 +198,7 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		newTest.setStatus(Status.START_TESTING);
 		try {
 			newTest.setStatus(Status.START_TESTING);
-			controller.saveOne(getTestUser(), newTest, false, model);
+			mockPerfTestController.saveOne(getTestUser(), newTest, false, model);
 			fail("test status id START_TESTING, can not be saved");
 		} catch (IllegalArgumentException e) {
 		}
@@ -205,7 +209,7 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 	public void testGetTestList() {
 		createPerfTest("new test1", Status.READY, new Date());
 		ModelMap model = new ModelMap();
-		controller.getAll(getTestUser(), null, null, null, new PageRequest(0, 10), model);
+		mockPerfTestController.getAll(getTestUser(), null, null, null, new PageRequest(0, 10), model);
 		Page<PerfTest> testPage = (Page<PerfTest>) model.get("testListPage");
 		List<PerfTest> testList = testPage.getContent();
 		assertThat(testList.size(), is(1));
@@ -223,7 +227,7 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		testAdmin.setTimeZone("Asia/Seoul");
 		testAdmin = userService.save(testAdmin);
 
-		controller.getAll(testAdmin, null, null, null, new PageRequest(0, 10), model);
+		mockPerfTestController.getAll(testAdmin, null, null, null, new PageRequest(0, 10), model);
 		@SuppressWarnings("unchecked")
 		Page<PerfTest> testPage = (Page<PerfTest>) model.get("testListPage");
 		List<PerfTest> testList = testPage.getContent();
@@ -249,7 +253,7 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		otherTestUser.setRole(Role.USER);
 		otherTestUser = userService.save(otherTestUser);
 		otherTestUser.setTimeZone("Asia/Seoul");
-		controller.getAll(otherTestUser, null, null, null, new PageRequest(0, 10), model);
+		mockPerfTestController.getAll(otherTestUser, null, null, null, new PageRequest(0, 10), model);
 		@SuppressWarnings("unchecked")
 		Page<PerfTest> testPage = (Page<PerfTest>) model.get("testListPage");
 		List<PerfTest> testList = testPage.getContent();
@@ -259,7 +263,7 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		// test no permission for other user
 		model.clear();
 		try {
-			controller.getOne(otherTestUser, test.getId(), model);
+			mockPerfTestController.getOne(otherTestUser, test.getId(), model);
 			assertTrue(false);
 		} catch (NGrinderRuntimeException e) {
 			assertTrue(true);
@@ -276,12 +280,12 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 
 		Sort sort = new Sort("testName");
 		Pageable pageable = new PageRequest(0, 10, sort);
-		controller.getAll(getTestUser(), strangeName, null, null, pageable, model);
+		mockPerfTestController.getAll(getTestUser(), strangeName, null, null, pageable, model);
 		Page<PerfTest> testPage = (Page<PerfTest>) model.get("testListPage");
 		List<PerfTest> testList = testPage.getContent();
 		assertThat(testList.size(), is(1));
 
-		controller.getAll(getTestUser(), strangeName.substring(2, 10), null, null, new PageRequest(0, 10), model);
+		mockPerfTestController.getAll(getTestUser(), strangeName.substring(2, 10), null, null, new PageRequest(0, 10), model);
 		testPage = (Page<PerfTest>) model.get("testListPage");
 		testList = testPage.getContent();
 		assertThat(testList.size(), is(1));
@@ -292,23 +296,23 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		String testName = "test1";
 		PerfTest test = createPerfTest(testName, Status.FINISHED, new Date());
 		ModelMap model = new ModelMap();
-		controller.getReport(model, test.getId());
+		mockPerfTestController.getReport(model, test.getId());
 
 		model.clear();
-		controller.getPerfGraph(test.getId(), "TPS,mean_time(ms)", true, 0);
+		mockPerfTestApiController.getPerfGraph(test.getId(), "TPS,mean_time(ms)", true, 0);
 
 		model.clear();
-		controller.getReportSection(getTestUser(), model, test.getId(), 700);
+		mockPerfTestController.getReportSection(getTestUser(), model, test.getId(), 700);
 	}
 
 	@Test
 	public void testGetMonitorData() {
 		String testName = "test1";
 		PerfTest test = createPerfTest(testName, Status.FINISHED, new Date());
-		controller.getMonitorGraph(test.getId(), "127.0.0.1", 0);
+		mockPerfTestApiController.getMonitorGraph(test.getId(), "127.0.0.1", 0);
 
 		long testId = 123456L;
-		controller.getMonitorGraph(testId, "127.0.0.1", 700);
+		mockPerfTestApiController.getMonitorGraph(testId, "127.0.0.1", 700);
 	}
 
 	@Test
@@ -317,23 +321,23 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		PerfTest test = createPerfTest(testName, Status.FINISHED, new Date());
 		HttpServletResponse resp = new MockHttpServletResponse();
 		try {
-			controller.downloadCSV(getTestUser(), test.getId(), resp);
+			mockPerfTestController.downloadCSV(getTestUser(), test.getId(), resp);
 		} catch (IllegalStateException e) {
 			// the report file doesn't exist
 			assertTrue(true);
 		}
 		resp.reset();
-		controller.downloadLog(getTestUser(), test.getId(), "log", resp);
+		mockPerfTestController.downloadLog(getTestUser(), test.getId(), "log", resp);
 	}
 
 	@Test
-	public void testRefreshTestRunning() {
+	public void testRefreshTestRunning() throws IOException {
 		String testName = "test1";
 		// it is not a running test, can not test get statistic data.
 		PerfTest test = createPerfTest(testName, Status.TESTING, new Date());
 		test.setPort(11011);
 		try {
-			controller.refreshTestRunning(getTestUser(), test.getId());
+			mockPerfTestApiController.refreshTestRunning(getTestUser(), test.getId());
 		} catch (NullPointerException e) {
 			assertTrue(true);
 		}
@@ -347,16 +351,16 @@ public class PerfTestControllerTest extends AbstractPerfTestTransactionalTest im
 		PerfTest test2 = createPerfTest(testName2, Status.START_AGENTS, new Date());
 
 		String ids = test.getId() + "," + test2.getId();
-		HttpEntity<String> rtnJson = controller.getStatuses(getTestUser(), ids);
-		assertThat(rtnJson.getBody(), notNullValue());
+		Map<String, Object> rtnJson = mockPerfTestApiController.getStatuses(getTestUser(), ids);
+		assertThat(rtnJson, notNullValue());
 	}
 
 	@Test
 	public void testSearchTag() {
-		HttpEntity<String> rtn = controller.searchTag(getAdminUser(), "");
-		assertThat(rtn.getBody(), notNullValue());
-		rtn = controller.searchTag(getAdminUser(), "test");
-		assertThat(rtn.getBody(), notNullValue());
+		List<String> rtn = mockPerfTestApiController.searchTag(getAdminUser(), "");
+		assertThat(rtn, notNullValue());
+		rtn = mockPerfTestApiController.searchTag(getAdminUser(), "test");
+		assertThat(rtn, notNullValue());
 	}
 
 }
